requirements.txt
aiohttp==3.9.1
wakeonlan==3.1.0
pyyaml==6.0.1
asyncio-dgram==2.1.2
